thay đổi giá trị tính lương sang double

+++ 21127046_W7_BT1/CHUYENVIEN.h
@@ -10,7 +10,7 @@ public:
 	CHUYENVIEN(string, string, int, vector<string>, int);
 	CHUYENVIEN(const CHUYENVIEN&);
 	~CHUYENVIEN();
-	int tinhluong();
+	double tinhluong();
 	void input();
 	void output();
 };
